[DP-163] fix: 데이터 상품 자투리 구매시 가격 업데이트 오류 수정

+++ src/main/java/com/dapanda/member/dto/response/FindDataResponse.java
@@ -1,11 +1,8 @@
 package com.dapanda.member.dto.response;
 
 
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
+import java.math.BigDecimal;
+import lombok.*;
 
 @Getter
 @Builder(access = AccessLevel.PRIVATE)
@@ -13,9 +10,9 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class FindDataResponse {
 
-	private float data;
+	private BigDecimal data;
 
-	public static FindDataResponse of(float data) {
+	public static FindDataResponse of(BigDecimal data) {
 
 		return FindDataResponse.builder()
 				.data(data)
