+++ src/main/java/com/dapanda/product/repository/ProductCustomRepository.java
@@ -8,6 +8,7 @@ import com.dapanda.product.dto.response.*;
 import com.dapanda.product.entity.ItemType;
 import com.dapanda.product.entity.ProductSortOption;
 import com.dapanda.trade.dto.MobileDataScrap;
+import java.math.BigDecimal;
 import java.util.List;
 import org.springframework.stereotype.Repository;
 
@@ -15,7 +16,7 @@ import org.springframework.stereotype.Repository;
 public interface ProductCustomRepository {
 
 	public CursorPageResponse<MobileDataSummary> findMobileDataByCursor(Long cursorId, int size,
-			ProductSortOption productSortOption, Float dataAmount);
+			ProductSortOption productSortOption, BigDecimal dataAmount);
 
 	public CursorPageResponse<WifiSummary> findWifiByCursor(Long cursorId, int size,
 			ProductSortOption productSortOption, boolean isOpen, Double latitude, Double longitude);
@@ -26,11 +27,11 @@ public interface ProductCustomRepository {
 
 	public List<String> findWifiImages(Long wifiId);
 
-	public Float sumSoldMobileDataAmountByMemberId(Long memberId);
+	public BigDecimal sumSoldMobileDataAmountByMemberId(Long memberId);
 
 	List<ReadSellingProductResponse> findSellingProduct(ReadSellingProductRequest request);
 
-	List<MobileDataScrap> findMobileDataScrap(float dataAmount, Long memberId);
+	List<MobileDataScrap> findMobileDataScrap(BigDecimal dataAmount, Long memberId);
 
 	FindMarketPriceResponse findMarketPrice(ItemType itemType);
