+++ src/main/java/com/dapanda/trade/dto/request/ScrapPurchaseMobileDataRequest.java
@@ -3,12 +3,13 @@ package com.dapanda.trade.dto.request;
 import com.dapanda.trade.dto.MobileDataScrap;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
+import java.math.BigDecimal;
 import java.util.List;
 
 public record ScrapPurchaseMobileDataRequest(
 
 		@NotNull(message = "총 데이터양은 필수입니다.")
-		Float totalAmount,
+		BigDecimal totalAmount,
 
 		@NotNull(message = "총 가격은 필수입니다.")
 		@Min(1)
